♻️ Refactor code Exercício 02 - Implemente a Cifra de César para criptografar e descriptografar uma mensagem
